refactor(backend): move generating password to UsersService
