+++ packages/backend/src/utils.ts
@@ -8,18 +8,6 @@ export function parseUserForResponse(user: User) {
   return returnData;
 }
 
-export function generateRandomPassword(length: number): string {
-  const charset = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()_+-=[]{}|;:,.<>?';
-  const passwordArray = [];
-
-  for (let i = 0; i < length; i++) {
-    const randomIndex = Math.floor(Math.random() * charset.length);
-    passwordArray.push(charset[randomIndex]);
-  }
-
-  return passwordArray.join('');
-}
-
 export class ResponseBuilder<T> {
   private errorMessage: string | undefined;
   private dataToSend: T | undefined;
